requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.5.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1
asyncpg==0.29.0
greenlet==3.0.0

# Telegram
aiogram==3.4.1

# Cache & Queue
redis==5.0.1
cachetools==5.3.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Text Processing
pymorphy2==0.9.1
rapidfuzz==3.6.1
phonenumbers==8.13.31

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Utils
gunicorn==21.2.0
httpx==0.27.0
orjson==3.9.15
nltk==3.8.1
scikit-learn==1.4.1.post1
numpy>=1.26.4
sentry-sdk[fastapi]==1.40.6
aiohttp==3.9.5
aiofiles==23.2.1
apscheduler==3.10.4
jinja2==3.1.4

# Development
pytest==8.0.2
pytest-asyncio==0.23.8
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
mypy==1.8.0
ruff==0.3.0